chore: reorganise structure
