+++ impakt-backend/api/models/__init__.py
@@ -1,0 +1,3 @@
+from .tables import *
+from .base import *
+from .client import *
