+++ impakt-backend/api/models/tables/__init__.py
@@ -1,0 +1,2 @@
+from .company import *
+from .company_description import *
