+++ impakt-backend/api/models/tables/company.py
@@ -1,7 +1,10 @@
-from .. import Base, CRUDMixin
+from ..base import Base, CRUDMixin
 
 from sqlalchemy.orm import mapped_column, Mapped
 from sqlalchemy import Integer, String
+from ..client import make_session, engine
+
+session = make_session(engine)
 
 
 class Company(Base, CRUDMixin["Company"]):
@@ -11,6 +14,10 @@ class Company(Base, CRUDMixin["Company"]):
     name: Mapped[str] = mapped_column(String, nullable=False)
     stock_ticker: Mapped[str] = mapped_column(String, nullable=True)
     website: Mapped[str] = mapped_column(String, nullable=True)
+
+    @classmethod
+    async def test(cls):
+        return await cls.get(session, {"id": 1})
 
     def __repr__(self):
         return f"Company(id={self.id!r}, name={self.name!r}, stock_ticker={self.stock_ticker!r}, website={self.website!r})"
