--- impakt-backend/db/__init__.py
@@ -1,2 +1,0 @@
-from .base import *
-from .client import *
