--- impakt-backend/db/models/__init__.py
@@ -1,2 +1,0 @@
-from .company import *
-from .company_description import *
